+++ tests/services/pre-decorator.js
@@ -23,7 +23,6 @@ const GROUPS_HEADER_KEY = 'groups-header-key'
 const CLIENTTYPE_HEADER_KEY = 'clienttype-header-key'
 const BACKOFFICE_HEADER_KEY = 'backoffice-header-key'
 
-// eslint-disable-next-line require-await
 module.exports = customService(async function clientGroups(service) {
   function requestUnmodifiedHandler(request) {
     return request.leaveOriginalRequestUnmodified()