+++ tests/services/service-proxy.js
@@ -20,7 +20,6 @@ const customService = require('../../index')()
 
 const otherServiceName = 'other-service'
 
-// eslint-disable-next-line require-await
 module.exports = customService(async function clientGroups(service) {
   service.addRawCustomPlugin('POST', '/default', async function handler(req) {
     const { some } = req.body